requirements.txt
appnope==0.1.2
APScheduler==3.6.3
backcall==0.2.0
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==2.0.9
decorator==5.1.0
idna==3.3
ipython==7.30.1
jedi==0.18.1
matplotlib-inline==0.1.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.24
ptyprocess==0.7.0
Pygments==2.10.0
python-telegram-bot==13.9
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.26.0
six==1.16.0
tornado==6.1
traitlets==5.1.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
wcwidth==0.2.5
